updating for new schema
