+++ data_entry/char_schema_templates/schema_9092-321_1003891_a.csv
@@ -1,0 +1,8 @@
+name,nominal,usl,lsl,precision,spec_type,char_type,gauge_type
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
+DIM x,1.000,1.005,0.995,3,two_tailed,diameter,caliper
