+++ data_entry/data/characteristics.csv
@@ -1,14 +1,89 @@
-name,nominal,usl,lsl,precision,measured,specification_type_id,characteristic_type_id,employee_id,part_id,gauge_id
-DIM 1,2,2.005,1.995,3,1.998,two_tailed,diameter,5137,1,cal001
-DIM 2,1.5,1.502,1.498,3,1.502,two_tailed,distance,5137,1,cal001
-DIM 3,3,3.01,2.99,2,2.998,two_tailed,radius,5137,1,vsys
-DIM 4,0,0.005,0,3,0.003,upper_tailed,position,6752,1,cmm1
-DIM 5,35,36,34,0,35.391,two_tailed,angle,5137,1,cmm1
-DIM 6,0.55,0.551,0.549,3,0.549,two_tailed,distance,6752,1,cmm1
-DIM 7,0.25,0.251,0.249,3,0.25,two_tailed,distance,6752,1,cmm1
-DIM 8,2,2.005,1.995,3,1.996,two_tailed,diameter,5137,1,ind1
-DIM 1,11,12,10,0,,two_tailed,angle,6752,2,cmm1
-DIM 2,45,46,44,0,,two_tailed,angle,6752,2,cmm1
-DIM 3,60,61,59,0,,two_tailed,angle,5137,2,cmm1
-DIM 4,1,1.001,0.999,3,,two_tailed,radius,5137,2,vsys
-DIM 5,0,0.005,0,3,,upper_tailed,concentricity,6752,2,vsys
+name,nominal,usl,lsl,measured,precision,is_deviated,check_id,specification_type_id,characteristic_type_id,deviation_id,frequency_type_id,gauge_id
+DIM 1,8.314,8.319,8.309,8.996757667,3,FALSE,1,1,17,,10,3
+DIM 2,5.427,5.432,5.422,6.289119909,3,FALSE,1,1,12,,10,2
+DIM 3,4.862,4.867,4.857,5.250262141,3,FALSE,1,2,0,,5,4
+DIM 4,7.085,7.09,7.08,7.085,3,FALSE,1,3,1,,8,6
+DIM 1,0.824,0.829,0.819,1.520173359,3,FALSE,2,3,7,,0,6
+DIM 2,5.49,5.495,5.485,5.758910274,3,FALSE,2,1,15,,0,6
+DIM 3,6.876,6.881,6.871,7.530682909,3,FALSE,2,1,10,,7,6
+DIM 4,4.026,4.031,4.021,4.608116619,3,FALSE,2,3,14,,11,3
+DIM 1,2.094,2.099,2.089,2.845483969,3,FALSE,3,2,15,,4,4
+DIM 2,2.906,2.911,2.901,2.906,3,FALSE,3,2,3,,11,5
+DIM 3,4.001,4.006,3.996,4.001,3,FALSE,3,2,11,,8,6
+DIM 4,7.168,7.173,7.163,8.149727253,3,FALSE,3,3,0,,11,4
+DIM 1,8.728,8.733,8.723,8.728,3,TRUE,4,3,15,1,5,6
+DIM 2,10.59,10.595,10.585,10.59,3,FALSE,4,1,20,,7,6
+DIM 3,4.605,4.61,4.6,5.370672549,3,FALSE,4,3,6,,1,5
+DIM 4,8.061,8.066,8.056,8.259583997,3,FALSE,4,1,4,,6,3
+DIM 1,4.533,4.538,4.528,4.616140087,3,FALSE,5,1,15,,10,5
+DIM 2,3.798,3.803,3.793,3.798,3,FALSE,5,2,14,,5,2
+DIM 3,4.845,4.85,4.84,4.900507483,3,FALSE,5,2,15,,1,5
+DIM 4,6.622,6.627,6.617,6.867803951,3,FALSE,5,3,14,,11,4
+DIM 1,3.279,3.284,3.274,3.279,3,FALSE,6,2,7,,4,2
+DIM 2,2.057,2.062,2.052,2.715174734,3,FALSE,6,1,6,,4,2
+DIM 3,3.826,3.831,3.821,3.826,3,FALSE,6,2,16,,7,5
+DIM 4,0.31,0.315,0.305,0.31,3,FALSE,6,1,13,,3,4
+DIM 1,2.446,2.451,2.441,2.446,3,FALSE,7,2,9,,6,6
+DIM 2,9.738,9.743,9.733,9.738,3,FALSE,7,1,15,,1,3
+DIM 3,0.935,0.94,0.93,0.935,3,FALSE,7,2,2,,4,4
+DIM 4,3.764,3.769,3.759,,3,FALSE,7,2,2,,7,4
+DIM 1,9.298,9.303,9.293,,3,FALSE,8,3,7,,10,2
+DIM 2,8.812,8.817,8.807,9.603996421,3,FALSE,8,1,20,,0,4
+DIM 3,1.164,1.169,1.159,1.529536585,3,FALSE,8,1,7,,1,6
+DIM 4,9.146,9.151,9.141,9.146,3,FALSE,8,2,14,,0,3
+DIM 1,6.899,6.904,6.894,6.957213644,3,FALSE,9,1,12,,7,2
+DIM 2,3.802,3.807,3.797,3.802,3,FALSE,9,2,0,,7,3
+DIM 3,1.292,1.297,1.287,1.817348498,3,FALSE,9,3,7,,4,6
+DIM 4,4.459,4.464,4.454,4.459,3,FALSE,9,3,2,,0,6
+DIM 1,8.753,8.758,8.748,8.753,3,FALSE,10,3,14,,0,4
+DIM 2,1.892,1.897,1.887,1.942887653,3,FALSE,10,3,20,,10,5
+DIM 3,2.984,2.989,2.979,2.984,3,FALSE,10,1,20,,9,2
+DIM 4,1.843,1.848,1.838,1.843,3,FALSE,10,2,19,,8,2
+DIM 1,3.226,3.231,3.221,3.226,3,FALSE,11,2,3,,0,1
+DIM 2,2.415,2.42,2.41,2.415,3,FALSE,11,3,4,,2,4
+DIM 3,7.066,7.071,7.061,7.584476272,3,FALSE,11,3,4,,4,5
+DIM 4,7.54,7.545,7.535,7.54,3,FALSE,11,2,9,,0,6
+DIM 1,2.345,2.35,2.34,,3,FALSE,12,3,10,,7,6
+DIM 2,0.199,0.204,0.194,0.199,3,FALSE,12,1,13,,11,6
+DIM 3,10.608,10.613,10.603,10.608,3,FALSE,12,2,13,,5,4
+DIM 4,3.256,3.261,3.251,,3,FALSE,12,2,3,,1,1
+DIM 1,7.226,7.231,7.221,7.74777015,3,FALSE,13,2,13,,1,1
+DIM 2,8.894,8.899,8.889,8.894,3,TRUE,13,3,3,2,8,3
+DIM 3,0.204,0.209,0.199,0.204,3,FALSE,13,3,10,,6,5
+DIM 4,1.399,1.404,1.394,1.399,3,FALSE,13,2,8,,0,4
+DIM 1,7.701,7.706,7.696,7.857528374,3,FALSE,14,2,11,,5,2
+DIM 2,7.177,7.182,7.172,7.772713167,3,FALSE,14,2,13,,10,3
+DIM 3,2.329,2.334,2.324,2.329,3,FALSE,14,1,16,,11,3
+DIM 4,0.388,0.393,0.383,0.388,3,FALSE,14,1,17,,11,5
+DIM 1,10.687,10.692,10.682,10.687,3,FALSE,15,2,13,,5,3
+DIM 2,9.162,9.167,9.157,9.162,3,FALSE,15,2,5,,8,6
+DIM 3,2.572,2.577,2.567,3.332175106,3,FALSE,15,3,20,,6,4
+DIM 4,4.684,4.689,4.679,5.085189518,3,FALSE,15,3,16,,7,4
+DIM 1,7.765,7.77,7.76,7.765,3,FALSE,16,2,16,,11,4
+DIM 2,0.527,0.532,0.522,1.486440967,3,FALSE,16,2,6,,1,5
+DIM 3,9.938,9.943,9.933,10.80435005,3,FALSE,16,3,3,,7,4
+DIM 4,4.973,4.978,4.968,4.973,3,FALSE,16,1,0,,11,3
+DIM 1,8.506,8.511,8.501,8.506,3,FALSE,17,1,20,,2,2
+DIM 2,2.212,2.217,2.207,2.560787018,3,FALSE,17,1,11,,1,5
+DIM 3,8.771,8.776,8.766,8.771,3,FALSE,17,3,5,,4,3
+DIM 4,10.184,10.189,10.179,10.32177937,3,FALSE,17,1,3,,9,1
+DIM 1,10.71,10.715,10.705,,3,FALSE,18,1,14,,10,2
+DIM 2,0.048,0.053,0.043,0.048,3,FALSE,18,1,9,,0,3
+DIM 3,3.366,3.371,3.361,3.366,3,FALSE,18,2,14,,7,4
+DIM 4,8.853,8.858,8.848,9.323254362,3,FALSE,18,2,6,,1,1
+DIM 1,4.117,4.122,4.112,4.214204724,3,FALSE,19,1,11,,1,3
+DIM 2,6.197,6.202,6.192,,3,FALSE,19,1,1,,1,1
+DIM 3,6.357,6.362,6.352,,3,FALSE,19,3,11,,8,2
+DIM 4,7.378,7.383,7.373,7.59322378,3,FALSE,19,3,0,,11,6
+DIM 1,7.582,7.587,7.577,7.647083639,3,FALSE,20,3,17,,0,5
+DIM 2,2.38,2.385,2.375,2.38,3,FALSE,20,2,0,,9,5
+DIM 3,10.74,10.745,10.735,11.00677556,3,FALSE,20,2,6,,9,4
+DIM 4,6.303,6.308,6.298,6.303,3,FALSE,20,3,12,,0,2
+DIM 1,1.052,1.057,1.047,1.429050745,3,TRUE,21,1,20,3,7,6
+DIM 2,1.812,1.817,1.807,1.812,3,TRUE,21,2,20,4,8,6
+DIM 3,9.9,9.905,9.895,10.79278474,3,FALSE,21,2,9,,6,3
+DIM 4,1.862,1.867,1.857,2.285643977,3,FALSE,21,1,18,,4,6
+DIM 1,9.902,9.907,9.897,10.73901675,3,FALSE,22,2,4,,11,2
+DIM 2,0.947,0.952,0.942,0.947,3,FALSE,22,1,18,,2,4
+DIM 3,6.726,6.731,6.721,,3,FALSE,22,1,8,,2,5
+DIM 4,7.44,7.445,7.435,7.44,3,FALSE,22,1,17,,9,3
