+++ data_entry/data/checks.csv
@@ -1,0 +1,23 @@
+part_index,datetime_measured,part_id,inspection_id,employee_id
+1,6/1/2021 08:00,4,1,4113
+2,6/1/2021 09:00,4,1,4113
+3,6/1/2021 10:00,4,1,4113
+4,6/1/2021 11:00,4,1,4113
+5,6/1/2021 12:00,5,1,4113
+6,6/1/2021 13:00,5,1,4113
+7,6/1/2021 14:00,5,1,4113
+1,4/20/2022 08:00,1,2,5397
+2,4/20/2022 09:00,1,2,5397
+3,4/20/2022 10:00,1,2,5397
+4,4/20/2022 11:00,1,2,5397
+5,4/20/2022 12:00,1,2,5137
+6,4/20/2022 13:00,1,2,5137
+7,4/20/2022 14:00,1,2,5137
+1,11/5/2023 08:00,2,3,6752
+2,11/5/2023 09:00,2,3,6752
+3,11/5/2023 10:00,2,3,6752
+4,11/5/2023 11:00,2,3,6752
+5,11/5/2023 12:00,2,3,4848
+1,8/28/2023 08:00,3,4,4848
+2,8/28/2023 09:00,3,4,4848
+3,8/28/2023 10:00,3,4,4848
