+++ data_entry/data/departments.csv
@@ -1,9 +1,9 @@
-id,description
-mfe,manufacturing engineering
-npd,new product development
-mkt,marketing
-cnc,cnc
-wld,welding
-fab,fabrication
-qc,quality control lab
-hr,human resources
+name,description
+MFE,manufacturing engineering
+NPD,new product development
+MKT,marketing
+CNC,cnc
+WLD,welding
+FAB,fabrication
+QC,quality control lab
+HR,human resources
