+++ data_entry/data/disposition_types.csv
@@ -1,5 +1,5 @@
-id
-pass
-fail
-deviated
-incomplete
+id,name
+0,Pass
+1,Fail
+2,Incomplete
+3,Deviated
