+++ data_entry/data/employee_projects.csv
@@ -1,6 +1,10 @@
 employee_id,project_id
-5137,p0000022587
-5137,c0000034255
-6752,p0000022587
-6752,c0000034255
-6752,m0003345279
+5137,1
+5137,2
+5137,3
+6752,1
+6752,3
+5397,2
+4848,1
+4848,2
+4848,3
