+++ data_entry/data/employees.csv
@@ -1,3 +1,9 @@
 id,first_name,last_name,department_id,location_id
-5137,john,doe,qc,ws10
-6752,alfred,waluigi,cnc,pad_25
+5137,John,Doe,1,2
+6752,Alfred,Waluigi,1,1
+8738,Simon,Johnson,3,5
+9689,Jane,Doe,7,6
+6981,Happy,Go-Lucky,8,3
+5397,Michael,Stevens,2,7
+4113,Jeffrey,Jefferson,5,7
+4848,Naomi,Nagata,5,7
