+++ data_entry/data/frequency_types.csv
@@ -1,0 +1,13 @@
+id,name
+0,1 / Shift
+1,"1 / Shift, Visual"
+2,Multiple / Shift
+3,First
+4,First & Last
+5,Every 10
+6,Every 20
+7,Every 30
+8,Every 40
+9,Every 50
+10,Every 100
+11,CMM
