+++ data_entry/data/gauges.csv
@@ -1,7 +1,7 @@
-id,last_calibrated,gauge_type_id,employee_id,location_id
-cal001,5/10/2022,caliper,5137,None
-cal002,6/2/2021,caliper,None,pad_32
-cmm1,6/1/2021,hexagon_cmm,None,qclab
-vsys,6/2/2021,vision_system,None,qclab
-ind1,6/2/2020,indicator,None,qclab
-dg001,2/2/2021,depth_gauge,None,qclab
+name,last_calibrated,gauge_type_id,employee_id,location_id
+CAL-001,2022/10/5,0,5137,
+CAL-002,2021/02/06,0,,6
+CMM-1,2021/01/06,4,,7
+VSYS-1,2021/02/06,5,,7
+IND-001,2020/02/06,3,,7
+DG-001,2021/02/02,2,,7
