+++ data_entry/data/inspection_lots.csv
@@ -1,0 +1,8 @@
+inspection_id,lot_id
+1,1
+1,2
+1,3
+4,1
+4,2
+2,4
+3,4
