+++ data_entry/data/inspection_purchase_orders.csv
@@ -1,6 +1,11 @@
 inspection_id,purchase_order_id
-0,SEN123456
-0,SEN846487
-1,SEN165984
-1,SEN123456
-1,SEN334484
+1,1
+1,2
+1,3
+2,1
+2,2
+1,5
+3,1
+3,2
+3,3
+3,4
