+++ data_entry/data/inspection_receiver_numbers.csv
@@ -1,6 +1,8 @@
 inspection_id,receiver_number_id
-0,RN1235584
-1,RN1235584
-0,RN3548611
-1,RN8944894
-1,RN1647684
+1,1
+1,2
+1,3
+4,1
+4,2
+2,5
+3,6
