+++ data_entry/data/inspection_reports.csv
@@ -1,3 +1,5 @@
-id,day_started,day_finished,full_inspect_qty,full_inspect_qty_type,released_qty,released_qty_type,completed_qty,completed_qty_type,material_type_id,supplier_id,job_order_id,employee_id,disposition,part_id
-0,1/1/2023,1/5/2023,0,first,30,custom,10,custom,steel,company_a,SEN0002345,5137,pass,1
-1,1/6/2023,1/8/2023,0,last,50,custom,22,custom,cfrp_composite,company_c,SEN0004567,6752,deviated,2
+full_inspect_interval,released_qty,completed_qty,material_type_id,supplier_id,job_order_id,employee_id,disposition_id
+500,1000,1000,0,1,1,5397,0
+575,1150,1150,2,2,5,6752,0
+400,800,800,2,3,6,4848,1
+88,176,85,1,4,3,4113,2
