+++ data_entry/data/locations.csv
@@ -1,8 +1,8 @@
-id,description,location_type_id
-ws10,workstation 10,work_station
-ws25,workstation 25,work_station
-pad_25,concrete pad 25,machine_pad
-wfh,working from home,working_remote
-wr,working remote,working_remote
-pad_32,concrete pad 32,machine_pad
-qclab,quality control lab,work_location
+name,description,location_type_id
+WS10,workstation 10,0
+WS25,workstation 25,0
+Pad 25,concrete pad 25,1
+WFH,working from home,3
+WR,working remote,3
+Pad 32,concrete pad 32,1
+QC Lab,quality control lab,5
