+++ data_entry/data/lots.csv
@@ -1,0 +1,5 @@
+name
+NHJ1234
+TGL8464
+CFO8891
+CVB6462
