+++ data_entry/data/machine_types.csv
@@ -1,5 +1,6 @@
-id
-cnc
-drill_press
-filament_winder
-label_laser
+id,name
+0,CNC
+1,Drill Press
+2,4-Axis Filament Winder
+3,1-Axis Filament Winder
+4,Label Laser
