+++ data_entry/data/material_types.csv
@@ -1,6 +1,6 @@
-id
-aluminum
-steel
-welded
-bent
-cfrp_composite
+id,name
+0,Aluminum
+1,Steel
+2,Welded
+3,Bent
+4,CFRP
