+++ data_entry/data/parts.csv
@@ -1,3 +1,7 @@
 drawing,revision,item
-1005656,a,9090-101
-7005894,c,1020-899
+2007247,a,9090-101
+8006944,c,1020-899
+2009768,b,9092-321
+7103276,a,7622-222
+7103276,b,7622-222
+8007627,d,6161-451
