+++ data_entry/data/project_types.csv
@@ -1,4 +1,4 @@
-id
-product_development
-continuous_improvement
-market_research
+id,name
+0,Product Development
+1,Continuous Development
+2,Market Research
