+++ data_entry/data/purchase_orders.csv
@@ -1,6 +1,6 @@
-id,supplier_id
-SEN123456,company_a
-SEN165984,company_d
-SEN846487,company_a
-SEN334484,company_c
-SEN339548,company_b
+name,supplier_id
+SEN123456,1
+SEN165984,4
+SEN846487,1
+SEN334484,3
+SEN339548,2
