--- data_entry/data/quantity_types.csv
@@ -1,5 +1,0 @@
-id
-first
-second
-last
-custom
