+++ data_entry/data/receiver_numbers.csv
@@ -1,4 +1,4 @@
-id
+name
 RN1235584
 RN3548611
 RN9985347
