+++ data_entry/data/suppliers.csv
@@ -1,5 +1,5 @@
-id
-company_a
-company_b
-company_c
-company_d
+name
+Company A
+Company B
+Company C
+Company D
